Fixes following bug in the landmark voting:
1. Take a picture of https://fi.wikipedia.org/wiki/Eiffel-torni#/media/Tiedosto:Tour_Eiffel_Wikimedia_Commons_(cropped).jpg
2. Vote "I think it's a landmark"

What happened:
App responded with "Not a landmark" even though Google recognized it as the Eiffel tower

Expected result:
App recognizes the landmark

Background:
CameraPage was setting the landmark results into its own local state. This meant that the VotingPage was unaware of the Google results. Fixed by setting the landmark result in the context.